requirements.txt
streamlit==1.32.0
openai-whisper==20231117
pandas==2.2.1
SpeechRecognition==3.10.1
pydub==0.25.1
